Moved firebase init to own file, called from app.vue
